Add ephemeral contacts snapshot for agent prompts
